+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - **Variant Buildings** Added new variants for buildings built by HytinyBuilds.
 - **Floating Gifts** Balloons carrying gifts will now float by the player occasionally and can be shot down for gold coins and special building variants.
 - **Tiered Building Staves** Added 5 tiers of building staff with increasing build radii. Built in the planning desk. Planning desk now has upgrades to unlock each tier of building staff.
+- **Replace Charter Command** Added /ah replace-charter command that respawns the charter where its supposed to be in cases where its been broken.
 
 ### Changed
 
@@ -18,6 +19,7 @@
 ### Fixed
 
 - **Villager Duplication Bug** Fixed bugs where multiples of a villager would spawn in the Inn.
+- **Charter Breaking** Charter would break if block under it was broken, this has been fixed.
 
 ## [1.3.0] - 5/4/2026